Remove floating point fiber test
The current implementation didn't accomplish anything. An inspection of
the decompiled asm reveales that everything is happening on the stack.
The calling convention is so deeply engrained into the compiler that it won't
even put data into the registers if it know it will just have to save them out
later. This is ideal from a performance standpoint, but not ideal for our tests.

Therefore, the only way to be totally sure that fiber switches are saving the 
registers properly is to manually write asm checking it. This is a work in 
progress on a different branch